pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aetheriaforge"
version = "0.1.0"
description = "EthereaLogic Databricks Suite — Intelligent Data Transformation Engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Anthony Johnson" },
]
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "scipy>=1.11",
]

[dependency-groups]
app = [
    "gradio>=6.10.0,<7",
    "plotly>=5.0,<6",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.0",
    "types-PyYAML>=6.0",
]
databricks = [
    "databricks-sdk>=0.20",
]
docs = [
    "matplotlib>=3.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aetheriaforge"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "aetheriaforge/templates"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["notebooks", "app"]

[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true
